+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## TBD
 ### Added
+- [#2042](https://github.com/org-roam/org-roam/pull/2042) Add `org-roam-db-extra-links-elements` and `org-roam-db-extra-links-exclude-keys` for fine-grained control over additional link parsing.
 ### Removed
 ### Fixed
 ### Changed